{{cookiecutter.repo_name}}/requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
Sphinx

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.2.2

# Django Extensions and relevant dependency
django-extensions==1.4.8
Werkzeug==0.9.6
